RA-18-Implement_Favorites_Screen

+++ app/src/main/java/com/practicum/androidsprint/Recipes.kt
@@ -242,7 +242,7 @@ object STUB {
     }
 
     fun getRecipesByCategoryId(categoryId: Int): List<Recipe> {
-        return if (categoryId.toInt() == 0) {
+        return if (categoryId == 0) {
             burgerRecipes
         } else listOf()
     }
